requirements.txt
flask
httpx
beautifulsoup4
PyPDF2
faiss-cpu
numpy
openai
python-dotenv